requirements-win.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
backoff==2.2.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.7
cycler==0.12.1
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
dnspython==2.8.0
elevenlabs==2.43.0
email-validator==2.3.0
fastapi==0.136.0
fastapi-cli==0.0.24
fastapi-cloud-cli==0.17.0
fastar==0.11.0
filelock==3.28.0
filetype==1.2.0
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.3.0
google-auth==2.49.2
google-genai==1.73.1
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.11.0
idna==3.7
imageio-ffmpeg==0.6.0
inference-sdk==1.2.3
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
llvmlite==0.47.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
networkx==3.6.1
numba==0.65.0
numpy==2.3.5
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==26.1
pi_heif==1.3.0
pillow==11.3.0
pillow-avif-plugin==1.5.5
propcache==0.4.1
protobuf==7.34.1
py-cpuinfo==9.0.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.2
pydantic-extra-types==2.11.1
pydantic-settings==2.13.1
pydantic_core==2.46.2
pyDeprecate==0.7.0
Pygments==2.20.0
pymongo==4.16.0
pynndescent==0.6.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.26
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
rich==15.0.0
rich-toolkit==0.19.7
rignore==0.7.6
roboflow==1.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentencepiece==0.2.1
sentry-sdk==2.58.0
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sports @ git+https://github.com/roboflow/sports.git@415751384e34f1cb7089a00909e0d7932fe73a75
starlette==1.0.0
supervision @ git+https://github.com/roboflow/supervision.git@6beabe3fb451e4709da05ba4911fbceea0911522
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==5.5.4
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
umap-learn==0.5.12
urllib3==2.6.3
uvicorn==0.44.0
watchfiles==1.1.1
websockets==16.0
yarl==1.23.0
